Breaking change: rename folder datasource to datasources
